Add support for multiple stacks and split dist between master and dev
Compiled assets are now stored in a folder structure matching
/$BRANCH/$STACK in the S3 bucket.
The stack distinction allows to support both cedar and cedar-14 (as
cedar is soon unsupported, new versions will not be available there).
The branch distinction allows to have different versions of PHP
extensions in the develop branch than in master, allowing to try newer
versions of extensions in develop (while it is possible only for PHP
versions currently).